requirements.txt
langchain
langchain_community
langchain_openai
python-dotenv
langchain-google-genai
